remove unnecessary folders and files
Remove all unnecessary folders and files.
Modify the file name of main file and 2 dictionary files. There are 4
files for using now:
*iKnow.py
*position_detected_keywords.txt
*taipeiLocationDict.txt
*yelp_tags_data.txt
