--- v1.0/OK_backup_20160103/position_detected_keywords
@@ -1,5 +1,0 @@
-附近
-裡
-上
-旁邊
-有
