--- v1.0/OK_backup_20160103/speech.txt
@@ -1,1 +1,0 @@
-我要到台大附近吃飯不想吃麵台大附近有什麼好吃的可以推薦希望明天不要再下雨露天
