--- v1.0/convenience
@@ -1,116 +1,0 @@
-convenience
-icecream
-coffee
-beer_and_wine
-shavedice
-organic_stores
-juicebars
-cakeshop
-delicatessen
-gourmet
-bubbletea
-desserts
-diyfood
-tea
-streetvendors
-grocery
-farmersmarket
-bakeries
-
-sandwiches
-gluten_free
-chinese
-mideastern
-asianfusion
-ethiopian
-russian
-bulgarian
-vegan
-caribbean
-hungarian
-scandinavian
-indonesian
-indpak
-cuban
-taiwanese
-cafes
-himalayan
-international
-turkish
-mediterranean
-mexican
-sushi
-hawaiian
-austrian
-bangladeshi
-laotian
-bistros
-pakistani
-basque
-brazilian
-greek
-german
-mex
-hotdogs
-irish
-latin
-moroccan
-foodstands
-syrian
-srilankan
-singaporean
-raw_food
-japanese
-breakfast_brunch
-cambodian
-belgian
-kebab
-salad
-french
-creperies
-brasseries
-persian
-polish
-thai
-seafood
-halal
-soup
-burgers
-australian
-hotpot
-fishnchips
-ukrainian
-delis
-hotdog
-bbq
-steak
-kosher
-modern_european
-peruvian
-vegetarian
-burmese
-cajun
-tradamerican
-italian
-buffets
-cafeteria
-british
-gastropubs
-filipino
-portuguese
-mongolian
-pizza
-tapasmallplates
-tapas
-spanish
-diners
-fondue
-vietnamese
-afghani
-arabian
-argentine
-chicken_wings
-chickenshop
-african
-korean
-malaysian
