--- v1.0/gtts_tt.py
@@ -1,7 +1,0 @@
-import subprocess
-from gtts import gTTS
-
-audio_file = "hello.mp3"
-tts = gTTS(text="Hello World!", lang="en")
-tts.save(audio_file)
-return_code = subprocess.call(["afplay", audio_file])
