+++ v1.0/speech.txt
@@ -1,1 +1,1 @@
-請關機謝謝
+關機
